+++ examples/origami_slice_animation.py
@@ -153,11 +153,14 @@ def sort_particles_for_origami(positions, particle_ids, box_size):
 
     sorted_positions = np.zeros((n_particles, 3), dtype=np.float64)
 
+    # Ensure integer types to prevent float indexing errors
+    particle_ids = particle_ids.astype(np.int64)
+
     if id_ordering == 'x-major':
         sorted_positions[particle_ids - 1] = positions
     else:  # z-major (GADGET-4 default)
         for i in range(n_particles):
-            idx = particle_ids[i] - 1
+            idx = int(particle_ids[i]) - 1
             z = idx % grid_size
             y = (idx // grid_size) % grid_size
             x = idx // (grid_size * grid_size)
